Updated alphaSynth to latest C# based version Updated alphaSynth plugin and player sample to be compatible with newest alphaSynth Changed all integer based sizes and coordinates to floats due to rounding errors in size calculations

+++ Source/AlphaTab/Platform/ICanvas.cs
@@ -26,8 +26,8 @@ namespace AlphaTab.Platform
     /// </summary>
     public interface ICanvas : IPathCanvas
     {
-        int Width { get; set; }
-        int Height { get; set; }
+        float Width { get; set; }
+        float Height { get; set; }
 
         RenderingResources Resources { get; set; }
