+++ Source/AlphaTab/Rendering/Effects/FingeringEffectInfo.cs
@@ -37,9 +37,9 @@ namespace AlphaTab.Rendering.Effects
                    (note.RightHandFinger != Fingers.NoOrDead && note.RightHandFinger != Fingers.Unknown);
         }
 
-        public override int GetHeight(EffectBarRenderer renderer)
+        public override float GetHeight(EffectBarRenderer renderer)
         {
-            return _maxGlyphCount * (int)(20 * renderer.Scale);
+            return _maxGlyphCount * (20 * renderer.Scale);
         }
 
         public override EffectBarGlyphSizing SizingMode
