+++ Source/AlphaTab/Rendering/Effects/NoteEffectInfoBase.cs
@@ -43,7 +43,7 @@ namespace AlphaTab.Rendering.Effects
         protected abstract bool ShouldCreateGlyphForNote(EffectBarRenderer renderer, Note note);
 
         public abstract EffectBarGlyphSizing SizingMode { get; }
-        public abstract int GetHeight(EffectBarRenderer renderer);
+        public abstract float GetHeight(EffectBarRenderer renderer);
         public abstract EffectGlyph CreateNewGlyph(EffectBarRenderer renderer, Beat beat);
 
         public virtual bool CanExpand(EffectBarRenderer renderer, Beat @from, Beat to)
