+++ Source/AlphaTab/Rendering/Glyphs/DiamondNoteHeadGlyph.cs
@@ -21,7 +21,7 @@ namespace AlphaTab.Rendering.Glyphs
     {
         private readonly bool _isGrace;
 
-        public DiamondNoteHeadGlyph(int x, int y, bool isGrace)
+        public DiamondNoteHeadGlyph(float x, float y, bool isGrace)
             : base(x, y, isGrace ? NoteHeadGlyph.GraceScale : 1, MusicFontSymbol.NoteHarmonic)
         {
             _isGrace = isGrace;
@@ -29,7 +29,7 @@ namespace AlphaTab.Rendering.Glyphs
 
         public override void DoLayout()
         {
-            Width = (int)(9 * (_isGrace ? NoteHeadGlyph.GraceScale : 1) * Scale);
+            Width = 9 * (_isGrace ? NoteHeadGlyph.GraceScale : 1) * Scale;
         }
 
         public override bool CanScale
