+++ Source/AlphaTab/Rendering/Glyphs/FlatGlyph.cs
@@ -21,7 +21,7 @@ namespace AlphaTab.Rendering.Glyphs
     {
         private readonly bool _isGrace;
 
-        public FlatGlyph(int x, int y, bool isGrace = false)
+        public FlatGlyph(float x, float y, bool isGrace = false)
             : base(x, y, isGrace ? NoteHeadGlyph.GraceScale : 1, MusicFontSymbol.AccidentalFlat)
         {
             _isGrace = isGrace;
@@ -29,7 +29,7 @@ namespace AlphaTab.Rendering.Glyphs
 
         public override void DoLayout()
         {
-            Width = (int)(8 * (_isGrace ? NoteHeadGlyph.GraceScale : 1) * Scale);
+            Width = 8 * (_isGrace ? NoteHeadGlyph.GraceScale : 1) * Scale;
         }
 
         public override bool CanScale
